+++ src/index.ts
@@ -19,7 +19,6 @@ function buildEnvironment(files: List<{ name: string, content: string }>, baseEn
   }), baseEnvironment)
 }
 
-
 export * from './model'
 export * from './interpreter/runtimeModel'
 export {