+++ src/parser.ts
@@ -1,5 +1,4 @@
 import Parsimmon, { takeWhile, alt as alt_parser, index, lazy, makeSuccess, notFollowedBy, of, Parser, regex, seq, seqObj, string, whitespace, any, Index } from 'parsimmon'
-import { basename, dirname } from 'path'
 import unraw from 'unraw'
 import { BaseProblem, SourceIndex, Assignment as AssignmentNode, Body as BodyNode, Catch as CatchNode, Class as ClassNode, Describe as DescribeNode, Entity as EntityNode, Expression as ExpressionNode, Field as FieldNode, If as IfNode, Import as ImportNode, Literal as LiteralNode, Method as MethodNode, Mixin as MixinNode, Name, NamedArgument as NamedArgumentNode, New as NewNode, Node, Package as PackageNode, Parameter as ParameterNode, Program as ProgramNode, Reference as ReferenceNode, Return as ReturnNode, Self as SelfNode, Send as SendNode, Sentence as SentenceNode, Singleton as SingletonNode, Super as SuperNode, Test as TestNode, Throw as ThrowNode, Try as TryNode, Variable as VariableNode, SourceMap, Closure as ClosureNode, ParameterizedType as ParameterizedTypeNode, Level, LiteralValue, Annotation, is } from './model'
 import { List, mapObject, discriminate } from './extensions'
@@ -120,16 +119,12 @@ const node = <N extends Node, P>(constructor: new (payload: P) => N) => (parser:
 export const File = (fileName: string): Parser<PackageNode> => lazy(() =>
   obj({
     fileName: of(fileName),
-    name: of(basename(fileName).split('.')[0]),
+    name: of(fileName.split('.')[0].replaceAll('/', '.')),
     imports: Import.sepBy(_).skip(_),
-    members: Entity.sepBy(_),
+    members: Entity.or(entityError).sepBy(_),
   }).skip(_)
-    .map(filePackage => {
-      const dir = dirname(fileName)
-      return (dir === '.' ? [] : dir.split('/')).reduceRight((entity, name) =>
-        new PackageNode({ name, members:[entity] })
-      , new PackageNode(filePackage))
-    }))
+    .map(filePackage => new PackageNode(recover(filePackage)))
+)
 
 
 export const Import: Parser<ImportNode> = node(ImportNode)(() =>